hw: vfio: quirks: Split ATI quirks out
And provide stubs when VFIO_QUIRKS_ATI is not set.
This way we can potentially reduce the amount of
VFIO PCI quirks that are built in.
