+++ default-configs/pci.mak
@@ -47,3 +47,4 @@ CONFIG_VFIO_VGA=y
 CONFIG_VIRTIO_INPUT_PCI=y
 CONFIG_VFIO_PCI_QUIRKS=y
 CONFIG_VFIO_PCI_QUIRKS_IGD=y
+CONFIG_VFIO_PCI_QUIRKS_ATI=y
